backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-filter>=23.0
python-decouple>=3.8
django-cors-headers>=4.0.0
django-ckeditor-5>=0.2.0
Pillow>=10.0.0
django-environ>=0.11.0
django-ratelimit>=4.1.0
requests>=2.31.0
social-auth-app-django>=5.4.0
social-auth-core>=4.5.0
django-simple-captcha>=0.6.0
